Add files via upload
This C++ program implements a linked list and a function to duplicate the second half of the list. The create function adds nodes to the end of the linked list, and the c function calculates the length of the list. The func function duplicates the second half of the list by creating new nodes with the same values, and the display function prints the contents of the list. In the main function, a linked list is initially created, its length is displayed, and then the second half is duplicated and the updated list is displayed again.