Add files via upload
This C++ program defines a binary tree structure and implements a function, nodesatk, which prints nodes at a distance 'k' from a specified target node. The checktarget function is utilized to find the node containing the target value in the binary tree. The nodesatk function then recursively explores the left and right subtrees of the target node, printing nodes at a distance 'k-1' until the specified distance is reached. In the main function, a sample binary tree is created, and the nodesatk function is called to print nodes at a distance of 2 from the target node with the value 5.