Add files via upload
This C++ program defines a binary tree structure and implements functions to find the Lowest Common Ancestor (LCA) of two nodes and calculate the shortest distance between two nodes in a binary tree. The lca function determines the LCA of two nodes, and the findDistance function calculates the distance from a target node to the root. The shortestdistance function utilizes these functions to find the shortest distance between two specified nodes in the binary tree. In the main function, a binary tree is created, and the shortestdistance function is called to find the shortest distance between nodes with values 8 and 7.