Add files via upload
This C++ program defines a binary tree structure and implements a function, sumreplacement, that performs a sum replacement operation on each node. The inorder function is then used to display the original tree structure, followed by the application of the sumreplacement function, and another inorder display to showcase the modified tree where each node value is replaced with the sum of its original value and the values of its left and right subtrees.